Remove a heckload of code I wrote to solve the day.
Quite clever code, if I may say that myself, and the fact that it
didn't work is sad to say the least.

The code removed applies all kinds of smart tricks on the program in
order to simplify it mathematically, like:

- simple symbolic reduction
- reductions based on arithmetic identities
- range-based reduction (e.g. eql is always false when comparing
disjointed intervals)
- hypothesis-based simplifications
- etc…

After that I wasn't happy and I even tried to print the program as a
single equation, to no avail.

As it turned out, the problem was more a pen and paper thing :(